Revert "merge .util package to .dict"
This reverts commit 67ed0169733d3fcde47946f89dd6abac5fb31bce.
