+++ gradle/generation/nori.gradle
@@ -19,7 +19,7 @@
 
 def recompileDictionary(project, dictionaryName, Closure closure) {
   project.javaexec {
-    main = "org.apache.lucene.analysis.ko.dict.DictionaryBuilder"
+    main = "org.apache.lucene.analysis.ko.util.DictionaryBuilder"
     classpath = project.sourceSets.main.runtimeClasspath
 
     jvmArgs '-Xmx1G'
