+++ lucene/analysis/kuromoji/src/java/module-info.java
@@ -24,6 +24,7 @@ module org.apache.lucene.analysis.kuromoji {
   exports org.apache.lucene.analysis.ja.completion;
   exports org.apache.lucene.analysis.ja.dict;
   exports org.apache.lucene.analysis.ja.tokenattributes;
+  exports org.apache.lucene.analysis.ja.util;
 
   opens org.apache.lucene.analysis.ja to
       org.apache.lucene.core;
