+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/tokenattributes/PartOfSpeechAttributeImpl.java
@@ -17,7 +17,7 @@
 package org.apache.lucene.analysis.ja.tokenattributes;
 
 import org.apache.lucene.analysis.ja.Token;
-import org.apache.lucene.analysis.ja.dict.ToStringUtil;
+import org.apache.lucene.analysis.ja.util.ToStringUtil;
 import org.apache.lucene.util.AttributeImpl;
 import org.apache.lucene.util.AttributeReflector;
