+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/util/ToStringUtil.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.analysis.ja.dict;
+package org.apache.lucene.analysis.ja.util;
 
 import java.io.IOException;
 import java.util.HashMap;
 
 /** Utility class for english translations of morphological data, used only for debugging. */
-public final class ToStringUtil {
+public class ToStringUtil {
   // a translation map for parts of speech, only used for reflectWith
   private static final HashMap<String, String> posTranslations = new HashMap<>();
