+++ lucene/analysis/kuromoji/src/java/org/apache/lucene/analysis/ja/util/package-info.java
@@ -14,20 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.analysis.ja.dict;
 
-/** Dictionary constants */
-public final class DictionaryConstants {
-  /** Codec header of the dictionary file. */
-  public static final String DICT_HEADER = "kuromoji_dict";
-  /** Codec header of the dictionary mapping file. */
-  public static final String TARGETMAP_HEADER = "kuromoji_dict_map";
-  /** Codec header of the POS dictionary file. */
-  public static final String POSDICT_HEADER = "kuromoji_dict_pos";
-  /** Codec header of the connection costs. */
-  public static final String CONN_COSTS_HEADER = "kuromoji_cc";
-  /** Codec header of the character definition file. */
-  public static final String CHARDEF_HEADER = "kuromoji_cd";
-  /** Codec version of the binary dictionary */
-  public static final int VERSION = 1;
-}
+/** Kuromoji utility classes. */
+package org.apache.lucene.analysis.ja.util;
