+++ lucene/analysis/kuromoji/src/test/org/apache/lucene/analysis/ja/TestCSVUtil.java
@@ -14,9 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.analysis.ja.dict;
+package org.apache.lucene.analysis.ja;
 
 import java.io.IOException;
+import org.apache.lucene.analysis.ja.util.CSVUtil;
 import org.apache.lucene.tests.util.LuceneTestCase;
 
 /*
