+++ lucene/analysis/nori/src/java/module-info.java
@@ -23,6 +23,7 @@ module org.apache.lucene.analysis.nori {
   exports org.apache.lucene.analysis.ko;
   exports org.apache.lucene.analysis.ko.dict;
   exports org.apache.lucene.analysis.ko.tokenattributes;
+  exports org.apache.lucene.analysis.ko.util;
 
   provides org.apache.lucene.analysis.TokenizerFactory with
       org.apache.lucene.analysis.ko.KoreanTokenizerFactory;
