+++ lucene/analysis/nori/src/java/org/apache/lucene/analysis/ko/KoreanTokenizer.java
@@ -197,7 +197,7 @@ public final class KoreanTokenizer extends Tokenizer {
   /**
    * Create a new KoreanTokenizer supplying a custom system dictionary and unknown dictionary. This
    * constructor provides an entry point for users that want to construct custom language models
-   * that can be used as input to {@link org.apache.lucene.analysis.ko.dict.DictionaryBuilder}.
+   * that can be used as input to {@link org.apache.lucene.analysis.ko.util.DictionaryBuilder}.
    *
    * @param factory the AttributeFactory to use
    * @param systemDictionary a custom known token dictionary
