+++ lucene/analysis/nori/src/java/org/apache/lucene/analysis/ko/util/TokenInfoDictionaryBuilder.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.analysis.ja.dict;
+package org.apache.lucene.analysis.ko.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -28,18 +28,12 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.apache.lucene.analysis.ja.dict.DictionaryBuilder.DictionaryFormat;
 import org.apache.lucene.util.IntsRefBuilder;
 import org.apache.lucene.util.fst.FST;
 import org.apache.lucene.util.fst.FSTCompiler;
 import org.apache.lucene.util.fst.PositiveIntOutputs;
 
-/** */
-final class TokenInfoDictionaryBuilder {
-
-  private final String encoding;
-  private final Normalizer.Form normalForm;
-  private final DictionaryFormat format;
+class TokenInfoDictionaryBuilder {
 
   /**
    * Internal word id - incrementally assigned as entries are read and added. This will be byte
@@ -47,9 +41,10 @@ final class TokenInfoDictionaryBuilder {
    */
   private int offset = 0;
 
-  public TokenInfoDictionaryBuilder(
-      DictionaryFormat format, String encoding, boolean normalizeEntries) {
-    this.format = format;
+  private String encoding;
+  private Normalizer.Form normalForm;
+
+  TokenInfoDictionaryBuilder(String encoding, boolean normalizeEntries) {
     this.encoding = encoding;
     normalForm = normalizeEntries ? Normalizer.Form.NFKC : null;
   }
@@ -67,38 +62,35 @@ final class TokenInfoDictionaryBuilder {
 
   private TokenInfoDictionaryWriter buildDictionary(List<Path> csvFiles) throws IOException {
     TokenInfoDictionaryWriter dictionary = new TokenInfoDictionaryWriter(10 * 1024 * 1024);
-    Charset cs = Charset.forName(encoding);
     // all lines in the file
     List<String[]> lines = new ArrayList<>(400000);
     for (Path path : csvFiles) {
-      try (BufferedReader reader = Files.newBufferedReader(path, cs)) {
+      try (BufferedReader reader = Files.newBufferedReader(path, Charset.forName(encoding))) {
         String line;
         while ((line = reader.readLine()) != null) {
           String[] entry = CSVUtil.parse(line);
 
-          if (entry.length < 13) {
+          if (entry.length < 12) {
             throw new IllegalArgumentException(
-                "Entry in CSV is not valid (13 field values expected): " + line);
+                "Entry in CSV is not valid (12 field values expected): " + line);
           }
 
-          lines.add(formatEntry(entry));
-
+          // NFKC normalize dictionary entry
           if (normalForm != null) {
-            if (Normalizer.isNormalized(entry[0], normalForm)) {
-              continue;
-            }
             String[] normalizedEntry = new String[entry.length];
             for (int i = 0; i < entry.length; i++) {
               normalizedEntry[i] = Normalizer.normalize(entry[i], normalForm);
             }
-            lines.add(formatEntry(normalizedEntry));
+            lines.add(normalizedEntry);
+          } else {
+            lines.add(entry);
           }
         }
       }
     }
 
     // sort by term: we sorted the files already and use a stable sort.
-    lines.sort(Comparator.comparing(entry -> entry[0]));
+    lines.sort(Comparator.comparing(left -> left[0]));
 
     PositiveIntOutputs fstOutput = PositiveIntOutputs.getSingleton();
     FSTCompiler<Long> fstCompiler = new FSTCompiler<>(FST.INPUT_TYPE.BYTE2, fstOutput);
@@ -108,21 +100,24 @@ final class TokenInfoDictionaryBuilder {
 
     // build token info dictionary
     for (String[] entry : lines) {
+      String surfaceForm = entry[0].trim();
+      if (surfaceForm.isEmpty()) {
+        continue;
+      }
       int next = dictionary.put(entry);
 
       if (next == offset) {
         throw new IllegalStateException("Failed to process line: " + Arrays.toString(entry));
       }
 
-      String token = entry[0];
-      if (!token.equals(lastValue)) {
+      if (!surfaceForm.equals(lastValue)) {
         // new word to add to fst
         ord++;
-        lastValue = token;
-        scratch.grow(token.length());
-        scratch.setLength(token.length());
-        for (int i = 0; i < token.length(); i++) {
-          scratch.setIntAt(i, (int) token.charAt(i));
+        lastValue = surfaceForm;
+        scratch.grow(surfaceForm.length());
+        scratch.setLength(surfaceForm.length());
+        for (int i = 0; i < surfaceForm.length(); i++) {
+          scratch.setIntAt(i, surfaceForm.charAt(i));
         }
         fstCompiler.add(scratch.get(), ord);
       }
@@ -131,60 +126,5 @@ final class TokenInfoDictionaryBuilder {
     }
     dictionary.setFST(fstCompiler.compile());
     return dictionary;
-  }
-
-  /*
-   * IPADIC features
-   *
-   * 0   - surface
-   * 1   - left cost
-   * 2   - right cost
-   * 3   - word cost
-   * 4-9 - pos
-   * 10  - base form
-   * 11  - reading
-   * 12  - pronounciation
-   *
-   * UniDic features
-   *
-   * 0   - surface
-   * 1   - left cost
-   * 2   - right cost
-   * 3   - word cost
-   * 4-9 - pos
-   * 10  - base form reading
-   * 11  - base form
-   * 12  - surface form
-   * 13  - surface reading
-   */
-
-  private String[] formatEntry(String[] features) {
-    if (this.format == DictionaryFormat.IPADIC) {
-      return features;
-    } else {
-      String[] features2 = new String[13];
-      features2[0] = features[0];
-      features2[1] = features[1];
-      features2[2] = features[2];
-      features2[3] = features[3];
-      features2[4] = features[4];
-      features2[5] = features[5];
-      features2[6] = features[6];
-      features2[7] = features[7];
-      features2[8] = features[8];
-      features2[9] = features[9];
-      features2[10] = features[11];
-
-      // If the surface reading is non-existent, use surface form for reading and pronunciation.
-      // This happens with punctuation in UniDic and there are possibly other cases as well
-      if (features[13].length() == 0) {
-        features2[11] = features[0];
-        features2[12] = features[0];
-      } else {
-        features2[11] = features[13];
-        features2[12] = features[13];
-      }
-      return features2;
-    }
   }
 }
