+++ lucene/analysis/nori/src/java/org/apache/lucene/analysis/ko/util/UnknownDictionaryWriter.java
@@ -14,14 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.analysis.ko.dict;
+package org.apache.lucene.analysis.ko.util;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import org.apache.lucene.analysis.ko.dict.CharacterDefinition;
+import org.apache.lucene.analysis.ko.dict.DictionaryConstants;
+import org.apache.lucene.analysis.ko.dict.UnknownDictionary;
 import org.apache.lucene.analysis.morph.BinaryDictionaryWriter;
 import org.apache.lucene.analysis.morph.CharacterDefinitionWriter;
 
-final class UnknownDictionaryWriter extends BinaryDictionaryWriter<UnknownDictionary> {
+class UnknownDictionaryWriter extends BinaryDictionaryWriter<UnknownDictionary> {
 
   private final org.apache.lucene.analysis.morph.CharacterDefinitionWriter<CharacterDefinition>
       characterDefinition =
